requirements.txt
pyzmq==15.3.0
elasticsearch==2.3.0
uvloop==0.6.7